Some code cleanup
